+++ projects/documentation/content/migrations/2021-8-11.11tydata.cjs
@@ -1,0 +1,335 @@
+/*
+Copyright 2020 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
+module.exports = {
+    versions: [
+        {
+            name: '@iliad-ui/accordion',
+            version: '0.6.0',
+            dir: '../../components/accordion',
+        },
+        {
+            name: '@iliad-ui/action-bar',
+            version: '0.4.0',
+            dir: '../../components/action-bar',
+        },
+        {
+            name: '@iliad-ui/action-button',
+            version: '0.7.0',
+            dir: '../../components/action-button',
+        },
+        {
+            name: '@iliad-ui/action-group',
+            version: '0.6.0',
+            dir: '../../components/action-group',
+        },
+        {
+            name: '@iliad-ui/action-menu',
+            version: '0.13.0',
+            dir: '../../components/action-menu',
+        },
+        {
+            name: '@iliad-ui/asset',
+            version: '0.6.0',
+            dir: '../../components/asset',
+        },
+        {
+            name: '@iliad-ui/avatar',
+            version: '0.9.0',
+            dir: '../../components/avatar',
+        },
+        {
+            name: '@iliad-ui/banner',
+            version: '0.7.0',
+            dir: '../../components/banner',
+        },
+        {
+            name: '@iliad-ui/base',
+            version: '0.5.0',
+            dir: '../../components/base',
+        },
+        {
+            name: '@iliad-ui/bundle',
+            version: '0.23.0',
+            dir: '../../components/bundle',
+        },
+        {
+            name: '@iliad-ui/button-group',
+            version: '0.8.0',
+            dir: '../../components/button-group',
+        },
+        {
+            name: '@iliad-ui/button',
+            version: '0.16.0',
+            dir: '../../components/button',
+        },
+        {
+            name: '@iliad-ui/card',
+            version: '0.10.0',
+            dir: '../../components/card',
+        },
+        {
+            name: '@iliad-ui/checkbox',
+            version: '0.12.0',
+            dir: '../../components/checkbox',
+        },
+        {
+            name: '@iliad-ui/coachmark',
+            version: '0.8.0',
+            dir: '../../components/coachmark',
+        },
+        {
+            name: '@iliad-ui/color-area',
+            version: '0.4.0',
+            dir: '../../components/color-area',
+        },
+        {
+            name: '@iliad-ui/color-handle',
+            version: '0.3.0',
+            dir: '../../components/color-handle',
+        },
+        {
+            name: '@iliad-ui/color-loupe',
+            version: '0.3.0',
+            dir: '../../components/color-loupe',
+        },
+        {
+            name: '@iliad-ui/color-slider',
+            version: '0.3.0',
+            dir: '../../components/color-slider',
+        },
+        {
+            name: '@iliad-ui/color-wheel',
+            version: '0.3.0',
+            dir: '../../components/color-wheel',
+        },
+        {
+            name: '@iliad-ui/dialog',
+            version: '0.8.0',
+            dir: '../../components/dialog',
+        },
+        {
+            name: '@iliad-ui/divider',
+            version: '0.4.0',
+            dir: '../../components/divider',
+        },
+        {
+            name: '@iliad-ui/dropzone',
+            version: '0.9.0',
+            dir: '../../components/dropzone',
+        },
+        {
+            name: '@iliad-ui/field-group',
+            version: '0.5.0',
+            dir: '../../components/field-group',
+        },
+        {
+            name: '@iliad-ui/field-label',
+            version: '0.7.0',
+            dir: '../../components/field-label',
+        },
+        {
+            name: '@iliad-ui/icon',
+            version: '0.11.0',
+            dir: '../../components/icon',
+        },
+        {
+            name: '@iliad-ui/icons-ui',
+            version: '0.8.0',
+            dir: '../../components/icons-ui',
+        },
+        {
+            name: '@iliad-ui/icons-workflow',
+            version: '0.8.0',
+            dir: '../../components/icons-workflow',
+        },
+        {
+            name: '@iliad-ui/icons',
+            version: '0.8.0',
+            dir: '../../components/icons',
+        },
+        {
+            name: '@iliad-ui/iconset',
+            version: '0.6.0',
+            dir: '../../components/iconset',
+        },
+        {
+            name: '@iliad-ui/illustrated-message',
+            version: '0.8.0',
+            dir: '../../components/illustrated-message',
+        },
+        {
+            name: '@iliad-ui/link',
+            version: '0.11.0',
+            dir: '../../components/link',
+        },
+        {
+            name: '@iliad-ui/menu',
+            version: '0.11.0',
+            dir: '../../components/menu',
+        },
+        {
+            name: '@iliad-ui/meter',
+            version: '0.6.0',
+            dir: '../../components/meter',
+        },
+        {
+            name: '@iliad-ui/modal',
+            version: '0.5.0',
+            dir: '../../components/modal',
+        },
+        {
+            name: '@iliad-ui/number-field',
+            version: '0.3.0',
+            dir: '../../components/number-field',
+        },
+        {
+            name: '@iliad-ui/overlay',
+            version: '0.13.0',
+            dir: '../../components/overlay',
+        },
+        {
+            name: '@iliad-ui/picker',
+            version: '0.9.0',
+            dir: '../../components/picker',
+        },
+        {
+            name: '@iliad-ui/popover',
+            version: '0.11.0',
+            dir: '../../components/popover',
+        },
+        {
+            name: '@iliad-ui/progress-bar',
+            version: '0.5.0',
+            dir: '../../components/progress-bar',
+        },
+        {
+            name: '@iliad-ui/progress-circle',
+            version: '0.4.0',
+            dir: '../../components/progress-circle',
+        },
+        {
+            name: '@iliad-ui/quick-actions',
+            version: '0.6.0',
+            dir: '../../components/quick-actions',
+        },
+        {
+            name: '@iliad-ui/radio',
+            version: '0.9.0',
+            dir: '../../components/radio',
+        },
+        {
+            name: '@iliad-ui/search',
+            version: '0.10.0',
+            dir: '../../components/search',
+        },
+        {
+            name: '@iliad-ui/shared',
+            version: '0.13.0',
+            dir: '../../components/shared',
+        },
+        {
+            name: '@iliad-ui/sidenav',
+            version: '0.12.0',
+            dir: '../../components/sidenav',
+        },
+        {
+            name: '@iliad-ui/slider',
+            version: '0.12.0',
+            dir: '../../components/slider',
+        },
+        {
+            name: '@iliad-ui/split-button',
+            version: '0.7.0',
+            dir: '../../components/split-button',
+        },
+        {
+            name: '@iliad-ui/split-view',
+            version: '0.4.0',
+            dir: '../../components/split-view',
+        },
+        {
+            name: '@iliad-ui/status-light',
+            version: '0.10.0',
+            dir: '../../components/status-light',
+        },
+        {
+            name: '@iliad-ui/styles',
+            version: '0.11.0',
+            dir: '../../components/styles',
+        },
+        {
+            name: '@iliad-ui/switch',
+            version: '0.9.0',
+            dir: '../../components/switch',
+        },
+        {
+            name: '@iliad-ui/tabs',
+            version: '0.8.0',
+            dir: '../../components/tabs',
+        },
+        {
+            name: '@iliad-ui/tags',
+            version: '0.8.0',
+            dir: '../../components/tags',
+        },
+        {
+            name: '@iliad-ui/textfield',
+            version: '0.10.0',
+            dir: '../../components/textfield',
+        },
+        {
+            name: '@iliad-ui/theme',
+            version: '0.9.0',
+            dir: '../../components/theme',
+        },
+        {
+            name: '@iliad-ui/thumbnail',
+            version: '0.4.0',
+            dir: '../../components/thumbnail',
+        },
+        {
+            name: '@iliad-ui/toast',
+            version: '0.10.0',
+            dir: '../../components/toast',
+        },
+        {
+            name: '@iliad-ui/tooltip',
+            version: '0.10.0',
+            dir: '../../components/tooltip',
+        },
+        {
+            name: '@iliad-ui/top-nav',
+            version: '0.5.0',
+            dir: '../../components/top-nav',
+        },
+        {
+            name: '@iliad-ui/tray',
+            version: '0.2.0',
+            dir: '../../components/tray',
+        },
+        {
+            name: '@iliad-ui/underlay',
+            version: '0.8.0',
+            dir: '../../components/underlay',
+        },
+        {
+            name: '@iliad-ui/story-decorator',
+            version: '0.4.0',
+            dir: '/Users/wesjohns/Documents/repos/spectrum-web-components/projects/story-decorator',
+        },
+        {
+            name: '@iliad-ui/vrt-compare',
+            version: '0.1.0',
+            dir: '/Users/wesjohns/Documents/repos/spectrum-web-components/projects/vrt-compare',
+        },
+    ],
+};